Moved Notification.h functions to a cpp, updated notification text to be more clear, switched over to using a char * instead of std::string
